Questao6

Documentation

f1 :: [[String]] -> [(String, Int)]

Esta função recebe uma lista de listas de string unindo essas listas em uma só lista,

de seguida zipa de modo a formar um tuplo cujo o segundo elemento vai do 1 até ao valor

do comprimento da lista concateada

f :: [[String]] -> [(String,Int)]
f lq = zip (concat lq) [1..]

Exemplos de Utilização:

Expand

>>> f [["A", "B", "C"] []]
[("A", 1), ("B", 2), ("C", 3)]

>>> f [["A", "B", "C"] ["D", "E"]]
[("A, 1"), ("B, 2"), ("C, 3"), ("D, 4"), ("E, 5")]

f2 :: [String] -> Int -> [[(String, Int)]]

f2 :: [String] -> Int -> [[(String,Int)]]
f2 [] _ = []
f2 l n = let 
            (l1,l2) = splitAt n l
            l1' = zip l1 [1..]
        in l1' : f2 l2 n

Exemplos de Utilização:

Expand

>>> f2 ["A", "B", "C", "D", "E"] 2
[[("A", 1), ("B", 2)], [("C", 1), ("D", 2)], [("E", 1)]]

f3 :: [String] -> String

Esta função recebe uma lista de Strings em que é usada uma função auxiliar de modo a transformar a string à cabeça numa lista de String isolando a head e o last dessa lista transformando de volta numa string de seguida aplicando a recursiva. Volta se à função principal em que Transforma essa lista de strings em uma string cujo cada elemento da lista anterior está separado por &

f3 :: [String] -> String
f3 [] = []
f3 ln = unwords ( intersperse "&" ln’ )
    where 
        ln' = aux ln
        aux :: [String] -> [String]
        aux [] = []
        aux (n:ns) = 
            let 
                lp = words n
                n' = unwords [head lp, last lp]
            in n' : aux ns

Exemplos de Utilização:

Expand

>>> f3 ["Daniel Marcos Costa", "Custodio Grande Gay", "Andre Pingu Burro"]
"Daniel Costa & Custodio Gay & Andre Burro"

>>> f3 ["Daniel Marcos Costa"]
"Daniel Costa"

f4 :: String -> String

f4 :: String -> String
f4 [] = []
f4 s = unwords (li ++ [last ln]) 
    where 
        ln = words s 
        li = aux (init ln) 
        aux :: [String] -> [String] 
        aux [] = []
        aux (n:ns) = [head n] : aux ns  

Exemplos de Utilização:

Expand

>>> f4 "Daniel Marcos Costa"
"D M Costa"
>>> f4 "Daniel Costa"
"D Costa"

teste1 :: Test

f5 :: [String] -> [(String, String)]

f5 :: [String] -> [(String,String)] 
f5 ls = let 
            n = div (length ls) 2 
            (l1,l2) = splitAt n ls   
        in zip (l1 ++ [" -- "]) l2

Exemplos de Utilização:

Expand

>>> f5 ["A", "B", "C", "D", "E"]
[("A", "C"), ("B", "D"), ("--", "E")]

>>> f5 ["A", "B", "C", "D"]
[("A", "C"), ("B", "D")]

f6 :: String -> String

Esta função recebe uma String de seguida transforma essa lista em uma lista de Strings, de seguida cria um tuplo de listas cujo cada comprimento de cada lista é metade do comprimento da lista original. Seguidamente une essas duas listas sendo inserido no meio a string "##" por fim transforma essa lista fina em uma String

f6 :: String -> String 
f6 t = 
    let 
        lp = words t 
        (lp1,lp2) = splitAt (div (length lp) 2) lp 
    in unwords (lp1 ++ ["##"] ++ lp2) 

Exemplos de Utilização:

Expand

>>> f6 ""
"##"
>>> f6 "A B C"
"A ## B C"