Tarefa1

Description

O objetivo desta tarefa é implementar um mecanismo de geração de labirintos. Os inputs serão o seu comprimento, largura e seed (mecanismo usado de forma a haver aleatoriadade na estrutura do labirinto.)

Documentation

geraAleatorios :: Int -> Int -> [Int]

Given a seed returns a list of n integer randomly generated

nrAleatorio :: Int -> Int

Através de uma seed, apresenta um número aleatório.

subLista :: Int -> [a] -> [[a]]

Converte uma lista numa lista de listas de tamanho n.

convertePeca :: Int -> Piece

Apresenta peça consoante o número apresentado

converteCorredor :: [Int] -> Corridor

Converte uma lista de inteiros num corredor

converteMaze :: [[Int]] -> Maze

Converte listas de inteiros num Labirinto com paredes em volta

printCorridor :: Corridor -> String

Função auxiliar para execução do labririnto

paredes :: Int -> Corridor

Adiciona paredes ao labirinto-

halflab :: Maze -> Int

Define o meio do labirinto

insidelab :: Maze -> Maze

Seleciona uma determinada area-

addtunels :: Maze -> Maze

Adiciona os tuneis

auxaddtunels :: Maze -> Maze

Função auxiliar para adicionar os túneis

bobinsidelab :: Int -> Int -> Int -> Maze

A partir dos inputs (3 números inteiros), apresenta um maze com as respetivas restrições impostas por outras funções nela contidas.

bobcasafantasmas :: Int -> Maze

Constroi a casa dos fantasmas.

putcasaeven :: Maze -> Maze -> Maze

Coloca a casa dos fantasmas na posiçao correta (labirinto par).

putcasaodd :: Maze -> Maze -> Maze

Coloca a casa dos fantasmas na posiçao correta (labirinto impar).

casaF :: Maze -> Maze

Preview de como deverá ser a casa dos fantasmas.

casaFantasmas :: Int -> Maze
casaFantasmas x 
   |even x = [[Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao],
              [Chao,Parede, Parede, Parede, Chao, Chao, Parede, Parede, Parede,Chao], 
              [Chao,Parede, Chao, Chao, Chao, Chao, Chao, Chao, Parede,Chao],
              [Chao,Parede, Parede, Parede, Parede, Parede, Parede, Parede, Parede,Chao],
              [Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao]]
  |otherwise =  [[Chao,Chao,Chao,Chao,Chao,Chao,Chao,Chao,Chao,Chao,Chao],
                 [Chao,Parede, Parede, Parede, Chao, Chao, Chao, Parede, Parede, Parede,Chao],
                 [Chao,Parede, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Parede,Chao],
                 [Chao,Parede, Parede, Parede, Parede, Parede, Parede, Parede, Parede, Parede,Chao],
                 [Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao, Chao,Chao]]

Construçao final da casa dos fantasmas.

fim :: Int -> Int -> Int -> IO ()

Labirinto final (com casa dos fantasmas e tuneis).

generateMaze :: Int -> Int -> Int -> Maze

Labirinto final em forma de lista.