Tarefa2

Documentation

mz :: Maze

Na Tarefa 2, foi nos pedido inicialmente formularmos a movimentação do Pacman e mais tarde, as movimentações também de Ghosts quando são controlados por jogadores reais. | Em primeiro lugar, para a movimentação do Pacman foram definidas várias funções auxiliares que dividiam o labirinto em várias partes, faziam a verificação das peças existentes e reorganizavam | listas de jogadores de forma a serem compatíveis com o resultado final. Com o auxilio destas funções, as funções mais crucias com por exemplo a função "direçao" ditavam as possibilidades e | os limites da movimentação do Pacman. | Em segundo lugar, foi preciso adaptar a estratégia de movimentação para ser possível a jogabilidade também com Ghosts. Para isso, foram necessárias alterações em algumas funções tais como | a função "playertostate" para ser compatível com a receção de um Ghost. | Por fim, a Tarefa 2 apresenta um papel crucial na movimentação dos jogadores e por isso tem de estar incluida nas tarefas seguintes.

Abreviaturas para a execução mais facil do jogo.

ps :: [Player]

l :: Integer

et :: State

mez :: Maze

js :: [Player]

Lista de jogadores presentes em jogo.

pacman :: Player

ghost1 :: Player

ghost2 :: Player

ghost3 :: Player

ghost4 :: Player

changeOrientation :: Orientation -> Player -> Player

Muda a direção do jogador.

corridorsize :: Maze -> Int

Identifica o comprimento do labirinto.

halflabv2 :: Maze -> Int

Segunda versão da função halflab que identifica a linha do meio do labirinto.

whichPiece :: Coords -> Maze -> Piece

Atráves de umas determinadas coordenadas e de um labirinto, identifica o tipo de peça.

isGhost :: Piece -> Bool

Verfica se uma determinada peça é um fantasma.

isPacman :: Piece -> Bool

Verfica se uma determinada peça é o Pacman.

moveL :: State -> State

Comportamento do jogador perante a movimentação para a esquerda em modo Normal (Pacman).

moveR :: State -> State

Comportamento do jogador perante a movimentação para a direita em modo Normal (Pacman).

moveU :: State -> State

Comportamento do jogador perante a movimentação para a cima em modo Normal (Pacman).

moveD :: State -> State

Comportamento do jogador perante a movimentação para a baixo em modo Normal (Pacman).

moveLMega :: State -> State

Comportamento do jogador perante a movimentação para a esquerda em modo Mega (Pacman).

moveRMega :: State -> State

Comportamento do jogador perante a movimentação para a direita em modo Mega (Pacman).

moveUMega :: State -> State

Comportamento do jogador perante a movimentação para a cima em modo Mega (Pacman).

moveDMega :: State -> State

Comportamento do jogador perante a movimentação para a baixo em modo Mega (Pacman).

moveLG :: State -> State

Comportamento do jogador perante a movimentação para a esquerda em modo Alive (Ghost).

moveRG :: State -> State

Comportamento do jogador perante a movimentação para a direita em modo Alive (Ghost).

moveUG :: State -> State

Comportamento do jogador perante a movimentação para a cima em modo Alive (Ghost).

moveDG :: State -> State

Comportamento do jogador perante a movimentação para baixo em modo Alive (Ghost).

direcao :: Play -> State -> State

Comportamento do jogador perante os diferentes casos de mudanças de direção (estado Normal e Mega).

ghostbehavior :: [Player] -> [Player]

Apresenta os diferentes cenários perante a mudança de estado de um jogador, quer seja Pacman ou Ghost.

ghostToHouse :: State -> State

Perante a interação de um Pacman em estado Mega e um Fantasma em estado Dead, retorna o fantasma para o interior da casa dos fantasmas em estado Alive.

solve :: State -> State

Verifica se 2 jogadores colidem.

colidestate :: Player -> State -> State

Faz com que um jogador adote um estado usado por outros jogadores.

death :: Player -> Player -> Player

Função auxiliar para a função clearpac.

clearpac :: Player -> State -> State

Retira o Pacman de uma lista (função auxiliar).

playertostate :: State -> State -> State

Perante o modo em que o Pacman (Normal ou Mega), o Fantasma adota diferentes comportamentos.

changestate :: State -> Player

Revela o state de um jogador (função auxiliar).

addstate :: Play -> State -> State

Verifica jogador (função auxiliar).

forceplayer :: Player -> State -> State

Identifica todos os jogadores diferentes do primeiramente mencionado (função auxiliar).

forcedplay :: Play -> State -> State

Identifica os jogadores com estados diferentes do primeiramente mencionado.

remove :: State -> [Player]

Através de um estado, retira um jogador.

transformstate :: State -> State

Coloca os jogadores "trabalhados" anteriormente no labirinto.

movimento :: Play -> State -> State

Permite aos jogadores realizar movimentos com todas as exceções colocadas nas funções anteriores (contidas na função).

play :: Play -> State -> State

Função final que apresenta o resultado pretendido.