Tarefa3

Documentation

compactCorredor :: Corridor -> Instruction

O objetivo desta tarefa é converter um labirinto válido numa sequência de instruções de modo a recriá-lo num formato mais compacto para leitura. A equipa decidiu percorrer o labirinto linha a linha para elaborar esta tarefa. Primeiramente, cada linha do labirinto era tranformada numa instrução, assim foi, com esse objetivo que definimos a função compactCorredor. Para tornar esta tarefa não só correta mas também óptima, adotamos uma estratégia mais compacta e económica no número de instruções, tal como foi sugerido no guião da respetiva tarefa. Como tal, para evitar a ocorrência de peças iguais no mesmo corredor, criamos a função element, para testar se existiam padrões iguais. Em caso afirmativo, com a função sameCorridor, é possível indicar ao interpretador que aplique ao corredor x o mesmo formato aplicado ao corredor y. Colocando uma instrução: Repeat y. Por fim, a função final faz com que a função sameCorridor seja aplicada a todo o labirinto, corredor a corredor.

Transforma cada peça de um corredor em instrução

element :: Instruction -> Instruction -> Bool

Testa se as instruções são iguais. Se são iguais então dá verdade, de outra forma dá falso

sameCorridor :: Instruction -> Instructions -> Int -> Instructions

Usando a função element, quando as instruções são iguais, então repete. De outra forma cria uma nova instrução

sameCorridor' :: Instructions -> Int -> Instructions

compactMaze :: Maze -> Instructions

Função final na qual a função sameCorridor é aplicada a todos os corredores