Tarefa4

Documentation

defaultDelayTime :: Int

passTime :: Int -> State -> State

Introdução

Para definirmos a função jogaTodos, tivemos especial atenção às jogadas efetuadas pelo pacman no modo mega e pelos fantasmas.

Função que dá a passagem do tempo: se delta (tempo decorrido desde a última jogada) é menor quer 250 ms então não se verifica qualquer alteração no estado do jogador, caso contrário, se delta é maior ou igual a 250 ms então aplicamos a função jogaTodos.

player :: State -> [Player]

jogaTodos :: [Player] -> State -> Int -> State

Função que é aplicada a cada 250 ms e faz com que todos os jogadores se movimentem, atualizando o seu estado depois de cada jogada.

getNextMove :: Player -> Play

isGhostv2 :: Player -> Bool

isPacmanMega :: Player -> Bool

getVelocidade :: Player -> Double

velocidade :: Double -> Player -> State -> State