Tarefa5

Documentation

ghostPlay :: State -> [Play]

Relativamente a esta tarefa, o objetivo seria realizar as movimentações dos fantasmas no labirinto, partindo de dois cenários. | Em primeiro lugar, a movimentação dos fantasmas vivos é dada pela função "chasemode" que através de várias funções auxiliares, em particular, a função "moveGhost" que aborda uma grande escala de situaçõe que o fantasma se pode encontrar e sua melhor solução para uma perseguição mais eficaz ao Pacman. | Em segundo lugar, a movimentação dos fantasmas mortos suportada pela função "scattermode" constituida por várias funções nomeadamente a "deadG" que dita o modo de fuga do fantasma em circulo para melhor fuga do Pacman em modo Mega. | Por fim, a Tarefa está desenvolvida para um jogabilidade com menos dificuldade para o Pacman visto que o comportamento dos fantasmas não apresenta um críterio tão exigente.

Função final (pelos 2 modos do fantasma, )

getGhostMode :: State -> GhostMode

Demonstra o estado de um fantasma na lista de jogadores (Alive ou Dead).

isGhostv2 :: Player -> Bool

Verifica se um jogador é fantasma.

firstG :: Player -> State -> State

Coloca um dado jogador em primeiro lugar da lista de jogadores do State.

lastPac :: State -> State

Organiza a lista de jogadores de forma a o Pacman ser o último ele

deadG :: State -> Coords -> Play

Aborda os diferentes casos para a movimentação de um fantasma morto.

isPacmanv2 :: Player -> Bool

Verifica se um jogador é o Pacman.

pac :: [Player] -> Coords

Obtem as coordenadas do Pacman, dada uma lista de jogadores. (auxiliar)

getPacCoords :: State -> Coords

Obtem as coordenadas do Pacman, dada uma lista de jogadores.

rearmode :: State -> Int -> Play

Através da verificação do id do jogador, aplica as instruções da função que formúla as movimentações de fantasmas mortos.

newin :: Player -> State -> State

Adiciona um novo jogador a lista de jogadores.

priority :: State -> State

Reorganiza a lista de jogadores de forma ao Pacman estar na última posição.

scatterMode :: State -> Int -> Play

Função final para fantasmas mortos.

distance :: Player -> Coords -> Float

Calcula a distancia entre as coordenadas de um fantasma e umas determinadas coordenadas do mapa (futuras coordenadas do Pacman).

moveGhost :: State -> Coords -> Play

Analisa todos os cenários de movimentação do fantasma e, tendo em conta os vários obstáculos, toma a melhor decisão para uma chegada mais rápida ao Pacman.

chasemode :: State -> Int -> Play

Função final para fantasmas vivos.