Tarefa6

Description

A Tarefa 6 consiste na criação de um "bot" que simule a jogabilidade de um jogador normal com jogadas pré-definidas em código. | Para isso, começamos por criar um mecanismo de intruções para a jogabilidade do Pacman no labirinto através da função "awayfromG" que define várias situações possíveis que o Pacman possa encontrar | no labirinto, em particular, alguns casos em que um Ghost se encontra perto do Pacman, este localiza a "Food Big" mais próxima e tenta dirigir se para lá (mecanismo criado com as funções "isFoodBig" | e "closerFood"). | Por fim, a tarefa exigiu funcionalidades idênticas à Tarefa 5, onde foram necessárias definir várias funcionalidas de comportamentos do Pacman suficiente para uma jogabilidade básica mas um pouco mais | escassas para uma jogabilidade avançada.

Documentation

isFoodBig :: Maze -> Coords -> [Coords]

Verifica se uma peça é uma "Food Big" e aplica recursividade.

closerFood :: [Coords] -> Coords -> Coords

Determina a distância mais curta entre duas coordenas.

awayfromG :: State -> Coords -> Play

Casos possíveis para a movimentação do bot em todo o labirinto (modo Normal).

closetoG :: State -> Coords -> Play

Casos possíveis para a movimentação do bot em todo o labirinto (modo Mega).

isGhostv3 :: State -> [Coords]

Verifica numa lista de jogadores de um State se é Ghost e caso se confirma, retira as suas coordenadas.

bot :: Int -> State -> Maybe Play