TesteModelo2

Documentation

f :: Eq a => [a] -> a -> Maybe Int

A função f pega em uma lista e em um elemento presente na lista,

zippando a lista original com a lista das posições, criando um par de tuplos

E devolvendo a posição do elemento inserido

f :: Eq a => [a] -> a -> Maybe Int
f l x = procura x li
             where li = zip l [0..]
                   procura x [] = Nothing
                   procura x ((y,n):ys) | x==y = Just n
                                        | otherwise = procura x ys

Exemplos de Utilização:

Expand

>>> f [1, 2, 3, 4] 3
Just 2
>>> f  [1, 2, 3, 4] 5
Nothing