Tarefa1

Description

Introdução

Nesta primeira tarefa, foi-nos proposto gerar labirintos com o túnel, a casa dos fantasmas e a parede lateral incluídos através de uma seed gerada aleatoriamente.

Objetivos e estratégias

Para a realização da tarefa dividimo-la em 4 partes:

As funções auxiliares do Generator para a tarefa, maioritariamente de conversões;

As funções para gerar o túnel e a parede lateral;

As funções para gerar a casa dos fantasmas;

Juntar todas as partes na função final generateMaze.

Conclusão

Para testarmos as funções acabamos por fazer uma função auxiliar que nos dá a preview, sendo mais fácil e rápido identificar se elas conseguiam obter os resultados esperados e saber onde corrigir, caso necessário.

Ao implementar estas funções cumprimos o objetivo e conseguimos gerar labirintos válidos, isto é, com a faixa central a formar um túnel, com todos os corredores com o mesmo comprimento, com a casa dos fantasmas localizada no centro e rodeado de peças Wall (exceto no túnel).

Documentation

convertsPiece :: Int -> Piece

Converte um Integer para Piece.

A função apenas deve receber números entre 0 e 99.

generateRandoms :: Int -> Int -> [Int]

Dada uma seed dá uma lista de n inteiros gerados aleatoriamente.

numbersToPieces :: [Int] -> Corridor

Converte uma lista de inteiros numa lista de peças.

listToMaze :: Int -> Corridor -> Maze

Dado um inteiro e uma lista de peças devolve um Maze

display :: Maze -> IO ()

Função que nos dá a preview.

updownWall :: Int -> Corridor

Devolve as paredes de cima e de baixo.

updownmaze :: Maze -> Maze

Coloca as paredes de cima e de baixo no labirinto.

sideWall :: Maze -> Maze

Devolve as paredes da esquerda e da direita.

replace :: Int -> Corridor -> Maze -> Maze

Encontra e substitui o corredor central no labirinto.

toTunnel :: Corridor -> Corridor

Converte um corredor em um corredor com peças Empty nos 2 extremidades.

addTunnel :: Int -> Maze -> Maze

Distingue se o labirinto tem altura par ou ímpar.

Usa as auxiliares oddTunnel e evenTunnel.

evenTunnel :: Int -> Int -> Maze -> Maze

Se a altura for par muda os 2 corredores centrais para túnel.

Encontra os 2 corredores centrais e depois coloca as peças Empty nas extremidades de cada um.

oddTunnel :: Int -> Int -> Maze -> Maze

Se a altura for ímpar muda o corredor central para túnel.

Encontra o corredor central e depois coloca as peças Empty em cada extremidade.

addHouse :: Int -> Maze -> Maze

Constrói a casa dos fantasmas ímpar ou par de acordo com a largura do labirinto.

Usa a auxiliar find5middle e as que dão a estrutura da casa (evenHouse and oddHouse).

find5middle

Arguments

:: Int
acumulador

-> Int
altura do maze

-> Maze
maze

-> Maze
maze com a casa (5 corredores)

-> Maze

Seleciona os 5 corredores centrais para colocar a casa.

Encontra o corredor onde a casa deve começar e substitui corredor por corredor o labirinto original pelo labirinto com a casa

Usa a auxiliar placeHouse.

placeHouse

Arguments

:: Int
acumulador

-> Int
largura do maze

-> Corridor
corredor do maze

-> Corridor
corredor da casa

-> Corridor

Coloca o corredor da casa num corredor do labirinto original.

Encontra a posição no corredor onde a casa deve começar e substitui peça por peça o corredor original pelo corredor com a casa.

oddHouse :: Maze

Estrutura da casa ímpar.

evenHouse :: Maze

Estrutura da casa par.

generateMaze

Arguments

:: Int
largura do maze

-> Int
altura do maze

-> Int
seed

-> Maze

Função principal que gera o labirinto.