Tarefa2

Description

Introdução

Nesta tarefa foi-nos proposta a implementação da função play que faz as jogadas e move os jogadores.

Objetivos e estratégias

Aplicamos uma função auxiliar para descobrir a peça para onde o jogador se quer mover.

Assim pudemos especificar cada caso particular e casos mais gerais e lógicos como o facto de um jogador não se poder mover para uma parede.

Em relação aos casos particulares, especificamos os seguintes casos:

Comportamento do pacman ao comer as comidas (Ganha 5 pontos se comer comida pequena e se comer comida grande 10 pontos e altera o modo para Mega);

Comportamento dos fantasmas quando o Pacman se encontra no modo Mega (devem reduzir a sua velocidade para metade da velocidade inicial e alterar o seu modo para Dead);

Comportamento do jogador quando se encontra numa peça do túnel (deve-se mover para o lado oposto do corredor);

Ocorrências se o pacman colidir com um fantasma se estiver em modo Mega (o fantasma morre e renasce para o centro da casa dos fantasmas e o Pacman ganha 10 pontos ao comer o fantasma);

Substituição das peças por onde o pacman passa (as comidas devem desaparecer após o Pacman passar por elas).

Na 1ª fase do trabalho limitamos a função final apenas para jogadas do pacman.

Com a 2ª fase do trabalho fizemos várias alterações a esta tarefa, referidas no relatório da T4 e da T5.

Com a experimentação do jogo no Main percebemos também que o pacman conseguia entrar dentro da casa dos fantasmas, o que não era suposto, e acrescentamos essa restrição à nossa função.

Conclusão

Ao concluir a função play, apesar da tarefa compilar no terminal, os testes na codeboard fornecida não coincidiam com o esperado e tínhamos ainda cerca de 10 erros em 13 possíveis da primeira vez que a testamos.

Estes inúmeros erros deviam-se ao facto de termos trocado as coordenadas, erro que corrijimos logo de seguida e reduzimos para apenas 3 erros.

Estes 3 erros foram mais difíceis de identificar mas atingimos o nosso objetivo e conseguimos concluir a tarefa com sucesso, passando a todos os testes.

Documentation

pieceMaze :: Coords -> Maze -> Piece

Retorna a peça respetiva às coordenadas dadas no Maze.

Esta função utiliza duas funções auxiliares: findpiece and findcorr.

findpiece :: Coords -> Corridor -> Piece

Encontra a peça respetiva às coordenadas dadas no corredor

findcorr :: Coords -> Maze -> Corridor

Encontra o corredor no Maze com as coordenadas dadas

movePacman :: Player -> Orientation -> Maze -> Player

Recebe um Player, uma Orientação e um Maze e devolve o Pacman com a nova Orientação ou com novas Coordenadas.

Se a Orientação dada é a mesma que a do Player então avança para a respetiva posição.

Se a Orientação dada é diferente da sua então altera a sua Orientação.

ATENÇÂO : Se existe uma _Wall_ na posição que o Player pretende ir, não se move.

moveGhost :: Player -> Orientation -> Maze -> Player

A função recebe um Player, uma Orientação, um Maze e devolve o Ghost com uma nova Orientação ou novas Coordenadas.

megaPacg :: Player -> [Player] -> [Player]

megaPacp :: Player -> Player

Atualiza o tempo Mega se o Pacman está em modo Mega.

switchMouth :: Player -> Player

Função que muda o estado da boca do Pacman.

posGhost :: Player -> [Player] -> Player

Aplica as alterações quando o Pacman colide com um Ghost (mesmas Coordenadas).

Quando o Pacman them mais de 1 vida e o Ghost está em modo Alive o Pacman perde 1 vida.

Quando o Pacman tem 0 vidas e o modo do Ghost é Alive o Pacman muda o seu modo para Dying-

posFood :: Player -> Maze -> Player

Altera a pontuação do Pacman quando está nas mesmas Coordenadas que alguma Food.

Quando o Pacman está nas mesmas Coordenadas que uma Food Little Pacman ganha 1 ponto.

Quando o Pacman está nas mesmas Coordenadas que uma Food Big ganha 5 pontos.

Esta função usa a função: pieceMaze

posTunnel :: Player -> Maze -> Orientation -> Player

Move o Pacman para o outro lado do tunel quando entra no lado contrário no tunel.

replacePiece :: Player -> Maze -> Maze

Substitui a comida, quando o Pacman tem as mesmas Coordenadas que a comida, por um espaço.

Esta função usa 2 auxiliares funções: pieceMaze e replaceElemInMaze.

ghostDie :: Player -> [Player] -> Maze -> [Player]

Descreve para onde os Ghosts vão após serem comidos pelo Pacman.

Se o Ghost estiver no modo Dead então ele vai para a casa dos Fantasmas.

A casa varia caso a altura seja par ou ímpar.

Esta função usa 2 auxiliares funções: getmiddlecoordsodd e getmiddlecoordseven.

getmiddlecoordsodd :: Maze -> Coords

Encontra onde é o meio do Maze quando a altura do Maze é ímpar.

getmiddlecoordseven :: Maze -> Coords

Encontra onde se situa o meio do Maze quando a altura é par.

ghosts :: [Player] -> [Player]

Reúne todos os Ghosts da lista de Players.

pacman :: [Player] -> Player

Encontra o Pacman na lista de jogadores.

moveGhosts :: [Player] -> Orientation -> Maze -> Int -> [Player]

Função que recebe um id do Ghost e a lista de Players e apenas move o Ghost com o respetivo id.

play :: Play -> State -> State

Função principal que aplica uma Play dado um Move e um State.