Tarefa3

Description

Introdução

Nesta tarefa compactamos o labirinto, tornando-o numa lista de instruções de modo a facilitar a leitura.

Objetivos e estratégias

No início da tarefa pensamos em primeiro elaborar o que está dentro da Instruct, ou seja, colocar tudo em pares de inteiros e peças e só depois focar na parte do Repeat, que acontece quando há corredores iguais.

Em relação ao que está dentro da instrução Instruct :

Agrupamos as peças iguais e seguidas de modo a tornar o labirinto em pares em que a primeira componente é o nº de peças seguidas iguais e a segunda a respetiva peça;

Para a realização de função utilizamos 2 funções auxiliares:

Uma para transformar o corredor em formato (1,Peça);

Outra que verifica se a peça a seguir é igual à anterior e se isto acontecer soma as primeiras componentes dos pares e mantém a segunda;

Já em relação à instrução Repeat :

Pretendemos comparar os corredores dos labirintos um a um para saber se deveríamos usar o Repeat caso já existisse um igual e ainda saber o índice da primeira vez que o corredor aparecia.

Assim sendo, utilizamos uma função que recebia 2 mazes iguais e comparava cada corredor do 1º Maze com todos do segundo e guardava o índice no 2º Maze da primeira vez que aparecesse um corredor igual ao corredor dado;

Para esta função recorremos a uma função auxiliar que dado um corredor devolvia a posição do corredor no Maze.

Com as funções terminandas, pudemos então juntar as funções e aplicar as instruções na função instructMaze.

Na função principal utilizamos a função instructMaze para aplicar as instruções e as funções mazeToTuple e corridorIndex para transformar o Maze dado nos tuplos pretendidos inicialmente e obter os índices dos corredores.

Conclusão

Apesar desta tarefa ser rápida e mais simples em relação às 2 anteriores, foi bastante útil para simplificar os nossos testes e exemplos e tornar o nosso código mais "limpo".

Concluímos também que nos ajudou a perceber o quão facilita às vezes recorrer a funções auxiliares e acumuladores porque foi algo bastante usado ao longo da tarefa e que ajudou imenso.

Documentation

corridorToTuple :: Corridor -> [(Int, Piece)]

Converte um corredor para uma lista de tuplos (Integer,Piece).

Exemplo :

corridorToTuple [Wall,Empty,Wall,Wall,Food Big] = [(1,Wall),(1,Empty),(1,Wall),(1,Wall),(1,Food Big)]

mazeToTuple :: Maze -> [[(Int, Piece)]]

Converte um labirinto numa lista de corredor no tipo (Integer,Piece).

Utiliza as funções corridorToTuple and gatherPieces.

Devolve já um labirinto onde os corredores são listas de tuplos e as peças seguidas estão juntas.

gatherPieces :: [(Int, Piece)] -> [(Int, Piece)]

A função gatherPieces junta tuplos seguidos se a peça for a mesma.

Exemplo :

gatherPieces [(1,Wall),(1,Empty),(1,Wall),(1,Wall),(1,Food Big)] = [(1,Wall),(1,Empty),(2,Wall),(1,Food Big)].

corridorIndex :: Int -> Corridor -> Maze -> Int

A função corridorIndex recebe um corredor e um labirinto e devolve a posição do corredor no labirinto.

Usamos um acumulador para ajudar a contar a posição. O acumulador começa sempre em 0.

mazeIndex :: Int -> Maze -> Maze -> [Int]

A função mazeIndex recebe 2 labirintos e compara cada corredor do primeiro Maze com todos do segundo.

Quando o corredor do 1º Maze é igual ao do 2º, a função guarda o índice do corredor no 2º Maze e aplica a função ao próximo corredor do 1º Maze.

Como comparamos labirintos iguais, a função verifica qual é o índice da 1ª vez que cada corredor do 1º maze aparece no 2º e guarda os índices numa lista.

Na função voltamos a recorrer a um acumulador.

Exemplo :

mazeIndex 0 [[Wall,Wall],[Wall,Food Big],[Wall,Food Big],[Wall,Wall],[Empty,Empty]] [[Wall,Wall],[Wall,Food Big],[Wall,Food Big],[Wall,Wall],[Empty,Empty]] = [0,1,1,0,4]

instructMaze :: Int -> [Int] -> [[(Int, Piece)]] -> Instructions

Recebe uma lista de números e um labirinto em tuplos e retorna a lista de corredores com Instruct e Repeat.

Voltamos a usar um acumulador.

Exemplo :

instructMaze 0 [0,1,1,0,4] [[(2,Wall)],[(1,Wall),(1,Food Big)],[(1,Wall),(1,Food Big)],[(2,Wall)],[(2,Empty)]] = [Instruct [(2,),(1,o)],Repeat 1,Repeat 0,Instruct [(2, )]]

compactMaze :: Maze -> Instructions

Função principal que compacta o maze.

Recebe um Maze e retorna o Maze com as peças agrupadas em corredores com Instruct e Repeat.

Exemplo :

compactMaze [[Wall,Wall],[Wall,Food Big],[Wall,Food Big],[Wall,Wall],[Empty,Empty]] = [Instruct [(2,),(1,o)],Repeat 1,Repeat 0,Instruct [(2, )]]