Tarefa5

Description

Introdução

Na Tarefa 5 foi-nos atribuída a tarefa onde conseguíssemos implementar um comportamento eficiente nos Ghosts para que assim conseguíssemos tornar o jogo mais interativo.

Objetivos

Para começar esta tarefa debatemos sobre várias alternativas para que os Ghosts se conseguissem movimentar independentemente, e com uma dificuldade acrescida para o Pacman conseguir fugir.

Começamos por dividir esta tarefa em várias partes para utilizar um método mais eficiente.

Criar uma função na Tarefa2 para que assim os Ghosts se consigam mover e estender a função play para a movimentação dos Ghosts.

Fazer com que os Ghosts persigam o Pacman quando este esteja em modo Normal e os Ghosts estejam em modo Alive.

Os Ghosts fugirem do Pacman quando este está em modo Mega e eles estejam no modo Dead.

Última parte, juntar as duas funções.

Decisões

Para começar o trabalho alteramos a Tarefa2:

Criamos a função na qual os Ghosts também se movem, no entanto estes não podem ir para os Túneis.

Como os Ghosts não podem atravessar os Túneis removemos essa opção.

Aplicando por fim a função criada na função principal play.

Na segunda parte do nosso trabalho criámos uma função chaseMode :

O Ghost vai verificar se o Pacman está mais à esquerda ou à direita.

Caso o Pacman eseja mais à direita e não esteja nenhuma Wall à sua direita então ele irá mover-se para a direita. O mesmo acontece se o Pacman estiver mais à direita, o Ghost irá mover-se para a esquerda.

Se houver uma parede na direção na qual o Ghost pretende ir então ele irá ver se o Pacman está em cima ou em baixo.

Quando o Pacman estiver abaixo então o Ghost irá mover-se para baixo, por outro lado se o Pacman estiver mais em cima então mover-se-á para cima.

Se o Ghost não conseguir concluir a jogada acima então terá que ir para a direção oposta.

Quando tiver as mesmas coordenadas dos y's então irá na direção dos x's do Pacman, conseguindo assim caçar o Pacman.

Na próxima fase criamos a função scatterMode para que o Ghost consiga fugir do Pacman quando este está em modo Mega:

Tal como na função chaseMode o Ghost irá verificar as coordenadas do Pacman.

Caso o Pacman esteja à sua direita e o Ghost não possua nenhuma Wall à sua esquerda então irá mover-se para a esquerda, afastando-se assim do Pacman.

Por outro lado se o Pacman estiver à esquerda então o Pacman vai para a direita, se não houver nenhuma Wall à sua direita.

Nos dois casos caso haja Wall para a posição na qual o Ghost pretende ir então ele irá verificar se o Pacman está mais acima ou abaixo e, assim, tomar a decisão correta.

Se o Pacman estiver em cima o Ghost vai para baixo. Caso contrário irá para cima. Isto se não houver nehnuma Wall a impedir o movimento.

Caso haja alguma Wall então irá tomar a orientação oposta.

Na última parte juntamos as duas funções:

Para juntar as duas funções ordenamos as listas de Players, para que o Pacman fique sempre em primeiro da lista, comparando-o com os restantes Ghosts.

Na função chaseMode apenas estão os Ghosts de modo Alive, para que assim persigam o Pacman.

Na função scatterMode estão os Ghosts no modo Dead para que fujam quando estiverem nesse modo.

Conclusão

Nesta tarefa tivemos algumas dificuldades em escolher qual o melhor método para que os Ghosts persigam o Pacman, visto que existem muitas formas.

Um problema que houve foi quando o Ghost ao fugir fica preso na Wall, não conseguindo resolver esse problema.

Documentation

ghostPlay :: State -> [Play]

Função final que recebe uma State com as listas ordenadas de ambas as funções finais.

chaseMode :: State -> Int -> Play

Função final na qual o Ghost respetivo ao ID recebido produz uma jogada para perseguir o Pacman.

scatterMode :: State -> Int -> Play

Função final na qual o Ghost respetivo ao ID recebido produz uma jogada para fugir do Pacman.

chaseMode' :: State -> Int -> Play

Ordena a lista do State na função chaseMode.

scatterMode' :: State -> Int -> Play

Ordena a lista do State na função chaseMode.

scatterMode1 :: State -> [Play]

Aplica a função scatterMode' a todos os Ghosts na lista de Players do State.

scatterMode1' :: State -> [Play]

Usa a função scatterMode1 com a lista ordenada do State.

chaseMode1 :: State -> [Play]

Aplica a função chaseMode' a todos os Ghosts na lista de Players do State.

chaseMode1' :: State -> [Play]

Usa a função chaseMode1 com a lista ordenada do State.