Tarefa6

Description

Introdução

Na Tarefa 6, foi-nos atribuído fazer um Bot que conseguisse jogar Pacman automaticamente.

Ao longo do jogo o Bot vai reagir de maneiras diferentes dependendo do seu modo e também de vários aspetos.

Objetivos

No início desta tarefa debatemos sobre como iríamos interpretar o enunciado e escolher a melhor forma para que o Pacman conseguisse completar o jogo ainda vivo.

Ponderamos criar um bot mais primitivo em que apenas fugia. No entanto, desse modo, o jogo ficava demasiado simples, não tendo assim muita diversidade de jogadas e decisões.

Então tentamos fazer um Bot com um pouco mais de complexidade.

Para conseguir organizar a tarefa dividimo-la em partes:

Quando o Bot está em modo Mega, onde vai perseguir os Ghosts que estavam apenas no modo Dead.

Quando o Bot não possui nenhum Ghost por perto, podendo assim "caçar" as peças, nomeadamente "comida", e assim ganhar pontos e também ter uma maior diversidade de jogadas.

Quando um Ghost se aproxima do Bot, ele deve ter como seu principal objetivo "fugir" do Ghost mais próximo e assim continuar no jogo.

Última parte, juntar as duas funções.

Decisões

Na primeira fase criamos a função catchGhost que faz com que o Bot persiga os Ghosts.

O nosso Bot apenas irá perseguir os Ghosts que estão no modo Dead, através da função allDead.

Então o Bot irá comparar a sua posição com a do Ghost.

Caso o Ghost esteja à sua direita e se não houver nenhuma Wall á sua direita, ele irá movimentar-se para a direita, tentando assim obter a mesma coordenada que o Ghost.

O mesmo acontece quando o Ghost está à sua esquerda, caso não tenha nenhuma parede à sua esquerda então ele irá mover-se para a esquerda.

No entanto se tiver uma Wall na coordenada para a qual o Bot se pretende mover então ele averiguará se o Ghost está acima ou abaixo.

Dependendo da situação, estando o Ghost acima então ele irá mover-se para cima não tendo nenhuma Wall em cima, isto é, ele mover-se-á para cima quando a coordenada em cima esteja livre.

Por outro lado quando o Ghost está em baixo ele irá movimentar-se para baixo se na coordenada em baixo não esteja uma Wall.

Em ambas as situações caso haja uma parede então ele terá que optar por tomar a direção contrária, ou seja, se estiver uma Wall em cima irá para baixo, caso haja uma Wall em baixo então o Bot irá para cima.

Quando o Bot conseguir atingir a mesma coordenada dos y's então irá procurar obter a mesma coordenada dos x's do Ghost, movimentando-se para cima ou para baixo, dependendo da posição do Ghost.

Na fase seguindo optamos por criar uma função a escapeGhost que recebe apenas Ghosts no modo Alive, onde o Bot, dependendo da sua distância ao Ghost mais próximo, irá averiguar se pode procurar pela "comida" mais próxima ou fugir do Ghost.

O nosso Bot irá procurar pela comida mais próxima caso o Ghost esteja a uma distância radial de 6, ou seja, se o Ghost estiver distante do Bot, ele pode procurar a comida mais próxima com a ajuda da função __proxFood'__.

Caso a comida esteja á sua direita e não haja Wall irá mover-se para a direita. O mesmo acontece se a comida estiver à esquerda, ele mover-se-á para a esquerda.

Se tiver uma Wall na direção na qual o Bot se move quando procura comida, se a comida estiver acima e não houver uma Wall acima move-se para cima, senão move-se para baixo.

O mesmo acontece quando a comida está para baixo, se não estiver uma Wall ele move-se para baixo, por outro lado terá que se mover para cima.

Na terceira fase da Tarefa decidimos colocar quando o Pacman terá que "fugir" dos Ghosts, visto que este método de procura está bastante relacionado com a distância do Bot ao Ghost mais próximo.

Do mesmo modo que o Bot faz para caçar o Ghost, neste caso ele faz o contrário.

Caso o Bot esteja mais para a direita então irá mover-se para a direita até bater contra uma Wall.

Quando colidir com a Wall irá determinar se o Ghost está mais acima ou abaixo e irá decidir conforme a sua posição. Se o Ghost estiver em cima, mover-se-á para baixo, caso contrário mover-se-á para cima.

Quando colide com uma Wall em baixo ou em cima irá também determinar se o Ghost está mais à esquerda ou direita.

Quando estiver na situação acima e o Ghost estiver mais à esquerda então move-se para a direita, senão move-se para a esquerda.

Uma estratégia para tornar o jogo mais interativo e divertido, foi colocar o Bot a passar no túnel quando está a uma distância inferior a 5 do Ghost mais próximo.

Na última parte do Trabalho decidimos juntar as funções escapeGhost' e catchGhost' utilizando-as alternadamente.

Caso o Bot esteja em modo Mega utilizamos a função catchGhost'.

Por outro lado, se estiver em modo Normal utilizamos a função escapeGhost'

Conclusão

Nesta tarefa tivemos algumas dificuldades em escolher qual o melhor método a utilizar, visto que existem inúmeros métodos, uns melhores que os outros.

Na nossa situação pensamos que o nosso método, apesar de não ser o mais eficiente, consegue atingir os objetivos e jogar fluidamente.

Alguns dos nossos problemas nesta Tarefa também passaram pela decisão que o Bot deve tomar quando está num canto, ou quando o Bot está no lado oposto ao Ghost.

Documentation

escapeGhost :: State -> Maybe Play

Esta função faz com que o bot consiga escapar dos Ghosts atribuindo jogadas dependendo do seu State.

proxGhost :: [Player] -> Coords

A função determina de uma lista de Players qual o Ghost mais próximo do Pacman dando as suas coordenadas.

proxGhost' :: [Player] -> Coords

Chama a função proxGhost com a lista de Players ordenada.

distancePac :: Coords -> Coords -> Int

A função calcula a distância entre duas coordenadas ao quadrado.

distancePac' :: Coords -> Coords -> Float

Transforma o valor obtido na distancePac na distância entre as duas coordenadas.

coordsFood :: Maze -> Coords -> [Coords]

Obtém uma lista com todas as coordenadas das comidas no Maze.

coordsFood' :: Maze -> [Coords]

Utiliza o contador na função coordsFood.

proxFood :: Player -> [Coords] -> Coords

Determina qual a coordenada mais próxima do Pacman.

proxFood' :: Player -> Maze -> Coords

Obtém a coordenada de comida, mais próxima do Pacman.

proxGhostO :: [Player] -> Orientation

A função retorna a orientação do Ghost mais próximo do Pacman.

proxGhostO' :: [Player] -> Orientation

Tem o mesmo resultado que a proxGhostO mas com a lista de Player ordenada.

catchGhost :: State -> Coords -> Maybe Play

Esta função faz com que o bot consiga perseguir os Ghosts atribuindo jogadas dependendo do seu State.

allDead :: [Player] -> Bool

Verifica se todos os Ghosts da lista estão em modo Dead.

escapeGhost' :: State -> Maybe Play

Função final na qual o Pacman foge dos Ghosts recebendo uma lista ordenada de Players no State.

catchGhost' :: State -> Maybe Play

Função final na qual o Pacman persegue os Ghosts recebendo uma lista ordenada de Players no State.

bot' :: Int -> State -> Maybe Play

Aplica uma jogada ao bot dependendo do seu modo.

bot :: Int -> State -> Maybe Play

Função final do bot que recebe uma lista ordenada de Players no estado.